Removed truncation on Storyboard item text + System Note Shortcut
The system note shortcut allow users to rapidly get at the main Panel on
the System Page (assuming they have not deleted it).

This allows users to work with layouts maximized easier than before (you
used to have to do a bunch of manual sliding of things around to get to
the system layout).